move callbacks and speed monitor
